requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
gunicorn>=20.0.0
tensorflow>=2.13.0
numpy>=1.21.0
opencv-python-headless>=4.5.0
Pillow>=9.0.0
matplotlib>=3.5.0
python-multipart>=0.0.5
jinja2>=3.0.0
